Add pre-publish-check for changesets
